services/agenticwork-mcp-proxy/requirements.txt
fastapi>=0.104.0,<1.0.0
uvicorn>=0.24.0,<1.0.0
httpx>=0.25.0,<1.0.0
pyjwt[crypto]>=2.8.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.5.0,<3.0.0
cryptography>=41.0.0,<50.0.0
mcp>=1.0.0,<2.0.0
azure-identity>=1.15.0,<2.0.0
requests>=2.31.0,<3.0.0
msal>=1.24.0,<2.0.0
redis>=5.0.0,<6.0.0
boto3>=1.34.0,<2.0.0
botocore>=1.34.0,<2.0.0